requirements.txt
transformers==4.31.0  
accelerate  
tensorboard  
fire  
sentencepiece  
torch_geometric  
seaborn  
pycocotools  
torchvision  
einops  
prdc  
pytorch_fid  
tiktoken==0.4.0  
-e git+https://github.com/microsoft/DeepSpeed.git@olruwase/ds_3948#egg=deepspeed  
